replace more int precision clocks/sample rates with double fp ones
git-svn-id: https://svn.code.sf.net/p/pinmame/code/trunk@5203 6f51dfdd-75df-4273-8f3b-818b1e6aa8bf

+++ src/mame.h
@@ -115,7 +115,7 @@ struct RunningMachine
 	/* ----- audio-related information ----- */
 
 	/* the digital audio sample rate; 0 if sound is disabled. */
-	int						sample_rate;
+	double					sample_rate;
 
 	/* samples loaded from disk */
 	struct GameSamples *	samples;
