+++ src/sound/fmopl.h
@@ -40,7 +40,7 @@ typedef unsigned char (*OPL_PORTHANDLER_R)(int param);
 
 #if BUILD_YM3812
 
-int  YM3812Init(int num, double clock, int rate);
+int  YM3812Init(int num, double clock, double rate);
 void YM3812Shutdown(void);
 void YM3812ResetChip(int which);
 int  YM3812Write(int which, int a, int v);
@@ -64,7 +64,7 @@ void YM3812SetUpdateHandler(int which, OPL_UPDATEHANDLER UpdateHandler, int para
 ** 'clock' is the chip clock in Hz
 ** 'rate' is sampling rate
 */
-int  YM3526Init(int num, double clock, int rate);
+int  YM3526Init(int num, double clock, double rate);
 /* shutdown the YM3526 emulators*/
 void YM3526Shutdown(void);
 void YM3526ResetChip(int which);
@@ -96,7 +96,7 @@ void Y8950SetPortHandler(int which, OPL_PORTHANDLER_W PortHandler_w, OPL_PORTHAN
 void Y8950SetKeyboardHandler(int which, OPL_PORTHANDLER_W KeyboardHandler_w, OPL_PORTHANDLER_R KeyboardHandler_r, int param);
 void Y8950SetDeltaTMemory(int which, void * deltat_mem_ptr, int deltat_mem_size );
 
-int  Y8950Init (int num, double clock, int rate);
+int  Y8950Init (int num, double clock, double rate);
 void Y8950Shutdown (void);
 void Y8950ResetChip (int which);
 int  Y8950Write (int which, int a, int v);
